Implement full functionality for replacement for damaged or lost license application and fix bugs of not editing the rows correctly in data grid ciew control cause of in correct row index between the data source physical data and the shown data

+++ Drivers And Vehicles License Department ( DVLD )/People/frmPeopleManagement.cs
@@ -219,7 +219,7 @@ namespace DVLDPresentationLayer
 
             if (PersonInfo != null)
             {
-                frmAddEditPersonInfo frm = new frmAddEditPersonInfo(PersonInfo,Convert.ToInt16(dgvPeople.SelectedRows[0].Index));
+                frmAddEditPersonInfo frm = new frmAddEditPersonInfo(PersonInfo,Convert.ToInt32(dgvPeople.SelectedRows[0].Index));
                 frm.AfterSavingEditedInfo += _EditDataRowInDGV;
                 frm.ShowDialog();
             }
